SortingTest to Support Single and Multiple Run
Updated the SortingTest class to allow for both single and multiple executions of sorting tests. you can now choose to run the sorting algorithms once or multiple times for each dataset size. When multiple runs are selected, the class calculates and reports the average execution time. The update maintains the flexibility of running a single test when needed. Tests are capped at a dataset size of 25600 to the 20-second execution limit in the assignment